Add feature to add breadcrumb viewer and config messages
Added a new command 'AddCommand' in Commands.kt to allow players to add another player as a viewer of their breadcrumbs. Also added corresponding config messages to reflect the updates in the event of success or failure of such add operation. Included three new functions in Config.kt to fetch these messages. This feature was added to enhance user interaction capabilities within the game.